requirements.txt
asgiref==3.7.2
boto3==1.28.40
botocore==1.31.40
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.5
djangorestframework==3.14.0
fpdf==1.7.2
idna==3.6
jmespath==1.0.1
Pillow==10.0.0
PyMySQL==1.1.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
whitenoise==6.6.0
